changed _enablePin property from int to byte

+++ src/TC_Multiplexer.h
@@ -28,13 +28,13 @@ class Multiplexer {
 
         /**
          * Turn the mux on, if enable pin was set
-         * @return Whether the enable pin is set.
+         * @return Whether the enable pin is set (ie, whether this operation did anything)
          */
         boolean setEnabled(bool enabled) {
             if (_enablePin) {
                 digitalWrite(_enablePin, enabled ? LOW : HIGH);
             }
-            return _enablePin;
+            return _enablePin; 
         }
 
         /**         
@@ -48,7 +48,7 @@ class Multiplexer {
 
     private:
         byte _pins[4];
-        int _enablePin; 
+        byte _enablePin; 
 
         /**
          * Set pin modes and enable pin.
